refactor: 네이밍, 함수 구조 수정
