+++ src/apis/queryFunctions/Auth.ts
@@ -1,7 +1,8 @@
-import apiClient from '@/apis/index';
+import apiClient from '@/apis/queryFunctions/apiClient';
 import { LoginParams, RegisterParams } from '@/apis/types/Auth';
 
-export const register = (param: RegisterParams) =>
+export const postRegister = (param: RegisterParams) =>
   apiClient.post<RegisterParams>('/v1/auth/register', param);
 
-export const login = (param: LoginParams) => apiClient.post<LoginParams>('/v1/auth/login', param);
+export const postLogin = (param: LoginParams) =>
+  apiClient.post<LoginParams>('/v1/auth/login', param);
